refactor the project layout
