+++ .gitignore
@@ -15,4 +15,4 @@
 # vendor/
 
 # goland settings
-.idea
+.idea/
